drop support for Sprockets 2.x
