+++ .travis.yml
@@ -7,7 +7,6 @@ rvm:
   - 2.2.4
   - 2.3.0
 gemfile:
-  - gemfiles/rails_4_2_sprockets_rails_2_sprockets_2.gemfile
   - gemfiles/rails_4_2_sprockets_rails_2.gemfile
   - gemfiles/rails_4_2.gemfile
   - gemfiles/rails_5_0_sprockets_rails_2.gemfile
