+++ Appraisals
@@ -1,9 +1,3 @@
-appraise "rails-4-2-sprockets-rails-2-sprockets-2" do
-  gem "rails", "~> 4.2.0"
-  gem "sprockets-rails", "~> 2.0"
-  gem "sprockets", "~> 2.0"
-end
-
 appraise "rails-4-2-sprockets-rails-2" do
   gem "rails", "~> 4.2.0"
   gem "sprockets-rails", "~> 2.0"
