--- gemfiles/rails_4_2_sprockets_rails_2_sprockets_2.gemfile
@@ -1,9 +1,0 @@
-# This file was generated by Appraisal
-
-source "https://rubygems.org"
-
-gem "rails", "~> 4.2.0"
-gem "sprockets-rails", "~> 2.0"
-gem "sprockets", "~> 2.0"
-
-gemspec :path => "../"
