+++ lib/spritely.rb
@@ -1,11 +1,8 @@
 require 'sass'
 require 'sprockets/rails/version'
 require 'sprockets/railtie'
-require 'sprockets/version'
 require 'spritely/sass_functions'
 require 'spritely/sprockets/manifest'
-require 'spritely/adapters/sprockets_2'
-require 'spritely/adapters/sprockets_3'
 
 module Spritely
   def self.environment
@@ -24,15 +21,7 @@ module Spritely
     Pathname.new(File.join('app', 'assets', 'images', 'sprites'))
   end
 
-  def self.sprockets_version
-    Gem::Version.new(::Sprockets::VERSION).segments.first
-  end
-
   def self.sprockets_rails_version
     Gem::Version.new(::Sprockets::Rails::VERSION).segments.first
-  end
-
-  def self.sprockets_adapter
-    Adapters.const_get("Sprockets#{sprockets_version}").new
   end
 end
