+++ lib/spritely/sass_functions.rb
@@ -1,10 +1,11 @@
 require 'spritely/sprite_map'
+require 'spritely/sprockets/adapter'
 
 module Spritely
   module SassFunctions
     def spritely_map(glob, kwargs = {})
       SpriteMap.create(glob.value, kwargs).tap do |sprite_map|
-        Spritely.sprockets_adapter.reset_cache!(sprockets_environment, sprite_map.filename)
+        Spritely::Sprockets::Adapter.reset_cache!(sprockets_environment, sprite_map.filename)
         sprockets_context.depend_on(Spritely.directory)
         sprite_map.files.each do |file|
           sprockets_context.depend_on(file)
