+++ lib/spritely/sprockets/adapter.rb
@@ -1,7 +1,7 @@
 module Spritely
-  module Adapters
-    class Sprockets3
-      def reset_cache!(environment, filename)
+  module Sprockets
+    module Adapter
+      def self.reset_cache!(environment, filename)
         environment.instance_variable_get(:@uris).delete_if { |_, asset| asset.pathname == filename }
         environment.instance_variable_get(:@entries).delete(Spritely.directory.to_s)
         environment.instance_variable_get(:@stats).delete(filename.to_s)
