+++ spec/spec_helper.rb
@@ -12,8 +12,6 @@ RSpec.configure do |config|
     c.syntax = :expect
   end
 
-  config.include "Sprockets#{Spritely.sprockets_version}SassFunctionsHelpers".constantize, :sass_functions
-
   config.include RailsAppHelpers, :integration
   config.include GeneratorSpec::GeneratorExampleGroup, :generator
   config.around(:each, :integration) { |example| within_rails_app(&example) }
