+++ spec/spritely_spec.rb
@@ -32,10 +32,4 @@ describe Spritely do
   describe '.relative_folder_path' do
     its(:relative_folder_path) { should eq(Pathname.new('app/assets/images/sprites')) }
   end
-
-  describe '.sprockets_version' do
-    before { stub_const('Sprockets::VERSION', '1.0') }
-
-    its(:sprockets_version) { should eq(1) }
-  end
 end
